# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Lens App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Verse {
  id: number;
  surah: number;
  ayah: number;
  arabic: string;
  surahName: string;
}

export interface Surah {
  number: number;
  name: string;
  nameArabic: string;
  verseCount: number;
}

export interface LensInfo {
  name: string;
  description: string;
  translationCount: number;
}

export interface GlossaryEntry {
  term: string;
  translation: string;
  /** @nullable */
  notes?: string | null;
}

export interface LensRules {
  name: string;
  description: string;
  rules: string;
  glossary: GlossaryEntry[];
  updatedAt: string;
}

export interface SaveLensRulesBody {
  description: string;
  rules: string;
  glossary: GlossaryEntry[];
}

export interface TranslatedVerse {
  id: number;
  surah: number;
  ayah: number;
  arabic: string;
  surahName: string;
  lensName: string;
  translated: string;
  generatedAt: string;
}

export interface GenerateTranslationsBody {
  /** Specific surahs to translate; empty means all */
  surahNumbers?: number[];
  /** Override the AI model to use */
  model?: string;
  /** Use remote API instead of local Ollama */
  useRemoteApi?: boolean;
}

export interface GenerationResult {
  lensName: string;
  totalVerses: number;
  successCount: number;
  errorCount: number;
  results: TranslatedVerse[];
  message: string;
}

export interface AiStatus {
  ollamaAvailable: boolean;
  ollamaUrl: string;
  defaultModel: string;
  useRemoteApi: boolean;
  remoteApiConfigured: boolean;
}

export type GetVersesParams = {
  /**
   * Filter by surah number
   */
  surah?: number;
  /**
   * Max number of verses to return
   */
  limit?: number;
};

export type GetTranslationsParams = {
  /**
   * Filter by surah number
   */
  surah?: number;
};