# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Quran Lens App API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: quran
    description: Quran data operations
  - name: lenses
    description: Lens rules operations
  - name: translations
    description: AI-generated translations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /quran/verses:
    get:
      operationId: getVerses
      tags: [quran]
      summary: Get all Quran verses
      parameters:
        - in: query
          name: surah
          schema:
            type: integer
          description: Filter by surah number
        - in: query
          name: limit
          schema:
            type: integer
          description: Max number of verses to return
      responses:
        "200":
          description: List of verses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Verse"

  /quran/surahs:
    get:
      operationId: getSurahs
      tags: [quran]
      summary: Get list of surahs
      responses:
        "200":
          description: List of surahs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Surah"

  /lenses:
    get:
      operationId: getLenses
      tags: [lenses]
      summary: Get all available lenses
      responses:
        "200":
          description: List of lenses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LensInfo"

  /lenses/{lensName}:
    get:
      operationId: getLensRules
      tags: [lenses]
      summary: Get rules for a specific lens
      parameters:
        - in: path
          name: lensName
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Lens rules
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LensRules"
        "404":
          description: Lens not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: saveLensRules
      tags: [lenses]
      summary: Save or update lens rules
      parameters:
        - in: path
          name: lensName
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveLensRulesBody"
      responses:
        "200":
          description: Lens rules saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LensRules"

  /translations/{lensName}:
    get:
      operationId: getTranslations
      tags: [translations]
      summary: Get existing translations for a lens
      parameters:
        - in: path
          name: lensName
          required: true
          schema:
            type: string
        - in: query
          name: surah
          schema:
            type: integer
          description: Filter by surah number
      responses:
        "200":
          description: Translations list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TranslatedVerse"
        "404":
          description: No translations yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /translations/{lensName}/generate:
    post:
      operationId: generateTranslations
      tags: [translations]
      summary: Generate AI translations for a lens
      parameters:
        - in: path
          name: lensName
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateTranslationsBody"
      responses:
        "200":
          description: Generation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResult"

  /ai/status:
    get:
      operationId: getAiStatus
      tags: [translations]
      summary: Check AI backend status (Ollama or remote)
      responses:
        "200":
          description: AI status info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiStatus"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    Verse:
      type: object
      properties:
        id:
          type: integer
        surah:
          type: integer
        ayah:
          type: integer
        arabic:
          type: string
        surahName:
          type: string
      required:
        - id
        - surah
        - ayah
        - arabic
        - surahName

    Surah:
      type: object
      properties:
        number:
          type: integer
        name:
          type: string
        nameArabic:
          type: string
        verseCount:
          type: integer
      required:
        - number
        - name
        - nameArabic
        - verseCount

    LensInfo:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        translationCount:
          type: integer
      required:
        - name
        - description
        - translationCount

    LensRules:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        rules:
          type: string
        glossary:
          type: array
          items:
            $ref: "#/components/schemas/GlossaryEntry"
        updatedAt:
          type: string
      required:
        - name
        - description
        - rules
        - glossary
        - updatedAt

    GlossaryEntry:
      type: object
      properties:
        term:
          type: string
        translation:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - term
        - translation

    SaveLensRulesBody:
      type: object
      properties:
        description:
          type: string
        rules:
          type: string
        glossary:
          type: array
          items:
            $ref: "#/components/schemas/GlossaryEntry"
      required:
        - description
        - rules
        - glossary

    TranslatedVerse:
      type: object
      properties:
        id:
          type: integer
        surah:
          type: integer
        ayah:
          type: integer
        arabic:
          type: string
        surahName:
          type: string
        lensName:
          type: string
        translated:
          type: string
        generatedAt:
          type: string
      required:
        - id
        - surah
        - ayah
        - arabic
        - surahName
        - lensName
        - translated
        - generatedAt

    GenerateTranslationsBody:
      type: object
      properties:
        surahNumbers:
          type: array
          items:
            type: integer
          description: Specific surahs to translate; empty means all
        model:
          type: string
          description: Override the AI model to use
        useRemoteApi:
          type: boolean
          description: Use remote API instead of local Ollama
      required: []

    GenerationResult:
      type: object
      properties:
        lensName:
          type: string
        totalVerses:
          type: integer
        successCount:
          type: integer
        errorCount:
          type: integer
        results:
          type: array
          items:
            $ref: "#/components/schemas/TranslatedVerse"
        message:
          type: string
      required:
        - lensName
        - totalVerses
        - successCount
        - errorCount
        - results
        - message

    AiStatus:
      type: object
      properties:
        ollamaAvailable:
          type: boolean
        ollamaUrl:
          type: string
        defaultModel:
          type: string
        useRemoteApi:
          type: boolean
        remoteApiConfigured:
          type: boolean
      required:
        - ollamaAvailable
        - ollamaUrl
        - defaultModel
        - useRemoteApi
        - remoteApiConfigured