# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Lens App API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get all Quran verses
 */
export const GetVersesQueryParams = zod.object({
  surah: zod.coerce.number().optional().describe("Filter by surah number"),
  limit: zod.coerce
    .number()
    .optional()
    .describe("Max number of verses to return"),
});

export const GetVersesResponseItem = zod.object({
  id: zod.number(),
  surah: zod.number(),
  ayah: zod.number(),
  arabic: zod.string(),
  surahName: zod.string(),
});
export const GetVersesResponse = zod.array(GetVersesResponseItem);

/**
 * @summary Get list of surahs
 */
export const GetSurahsResponseItem = zod.object({
  number: zod.number(),
  name: zod.string(),
  nameArabic: zod.string(),
  verseCount: zod.number(),
});
export const GetSurahsResponse = zod.array(GetSurahsResponseItem);

/**
 * @summary Get all available lenses
 */
export const GetLensesResponseItem = zod.object({
  name: zod.string(),
  description: zod.string(),
  translationCount: zod.number(),
});
export const GetLensesResponse = zod.array(GetLensesResponseItem);

/**
 * @summary Get rules for a specific lens
 */
export const GetLensRulesParams = zod.object({
  lensName: zod.coerce.string(),
});

export const GetLensRulesResponse = zod.object({
  name: zod.string(),
  description: zod.string(),
  rules: zod.string(),
  glossary: zod.array(
    zod.object({
      term: zod.string(),
      translation: zod.string(),
      notes: zod.string().nullish(),
    }),
  ),
  updatedAt: zod.string(),
});

/**
 * @summary Save or update lens rules
 */
export const SaveLensRulesParams = zod.object({
  lensName: zod.coerce.string(),
});

export const SaveLensRulesBody = zod.object({
  description: zod.string(),
  rules: zod.string(),
  glossary: zod.array(
    zod.object({
      term: zod.string(),
      translation: zod.string(),
      notes: zod.string().nullish(),
    }),
  ),
});

export const SaveLensRulesResponse = zod.object({
  name: zod.string(),
  description: zod.string(),
  rules: zod.string(),
  glossary: zod.array(
    zod.object({
      term: zod.string(),
      translation: zod.string(),
      notes: zod.string().nullish(),
    }),
  ),
  updatedAt: zod.string(),
});

/**
 * @summary Get existing translations for a lens
 */
export const GetTranslationsParams = zod.object({
  lensName: zod.coerce.string(),
});

export const GetTranslationsQueryParams = zod.object({
  surah: zod.coerce.number().optional().describe("Filter by surah number"),
});

export const GetTranslationsResponseItem = zod.object({
  id: zod.number(),
  surah: zod.number(),
  ayah: zod.number(),
  arabic: zod.string(),
  surahName: zod.string(),
  lensName: zod.string(),
  translated: zod.string(),
  generatedAt: zod.string(),
});
export const GetTranslationsResponse = zod.array(GetTranslationsResponseItem);

/**
 * @summary Generate AI translations for a lens
 */
export const GenerateTranslationsParams = zod.object({
  lensName: zod.coerce.string(),
});

export const GenerateTranslationsBody = zod.object({
  surahNumbers: zod
    .array(zod.number())
    .optional()
    .describe("Specific surahs to translate; empty means all"),
  model: zod.string().optional().describe("Override the AI model to use"),
  useRemoteApi: zod
    .boolean()
    .optional()
    .describe("Use remote API instead of local Ollama"),
});

export const GenerateTranslationsResponse = zod.object({
  lensName: zod.string(),
  totalVerses: zod.number(),
  successCount: zod.number(),
  errorCount: zod.number(),
  results: zod.array(
    zod.object({
      id: zod.number(),
      surah: zod.number(),
      ayah: zod.number(),
      arabic: zod.string(),
      surahName: zod.string(),
      lensName: zod.string(),
      translated: zod.string(),
      generatedAt: zod.string(),
    }),
  ),
  message: zod.string(),
});

/**
 * @summary Check AI backend status (Ollama or remote)
 */
export const GetAiStatusResponse = zod.object({
  ollamaAvailable: zod.boolean(),
  ollamaUrl: zod.string(),
  defaultModel: zod.string(),
  useRemoteApi: zod.boolean(),
  remoteApiConfigured: zod.boolean(),
});