# lib/api-zod/src/generated/types/aiStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Lens App API
 * OpenAPI spec version: 0.1.0
 */

export interface AiStatus {
  ollamaAvailable: boolean;
  ollamaUrl: string;
  defaultModel: string;
  useRemoteApi: boolean;
  remoteApiConfigured: boolean;
}