# lib/api-zod/src/generated/types/generateTranslationsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Lens App API
 * OpenAPI spec version: 0.1.0
 */

export interface GenerateTranslationsBody {
  /** Specific surahs to translate; empty means all */
  surahNumbers?: number[];
  /** Override the AI model to use */
  model?: string;
  /** Use remote API instead of local Ollama */
  useRemoteApi?: boolean;
}