# lib/api-zod/src/generated/types/generationResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Lens App API
 * OpenAPI spec version: 0.1.0
 */
import type { TranslatedVerse } from "./translatedVerse";

export interface GenerationResult {
  lensName: string;
  totalVerses: number;
  successCount: number;
  errorCount: number;
  results: TranslatedVerse[];
  message: string;
}