# lib/api-zod/src/generated/types/glossaryEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Lens App API
 * OpenAPI spec version: 0.1.0
 */

export interface GlossaryEntry {
  term: string;
  translation: string;
  /** @nullable */
  notes?: string | null;
}