# lib/api-zod/src/generated/types/saveLensRulesBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Lens App API
 * OpenAPI spec version: 0.1.0
 */
import type { GlossaryEntry } from "./glossaryEntry";

export interface SaveLensRulesBody {
  description: string;
  rules: string;
  glossary: GlossaryEntry[];
}