# lib/api-zod/src/generated/types/surah.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Lens App API
 * OpenAPI spec version: 0.1.0
 */

export interface Surah {
  number: number;
  name: string;
  nameArabic: string;
  verseCount: number;
}