# lib/api-zod/src/generated/types/translatedVerse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Quran Lens App API
 * OpenAPI spec version: 0.1.0
 */

export interface TranslatedVerse {
  id: number;
  surah: number;
  ayah: number;
  arabic: string;
  surahName: string;
  lensName: string;
  translated: string;
  generatedAt: string;
}